Use async function to commmit offset
