+++ lib/index.js
@@ -3,10 +3,8 @@ function __export(m) {
     for (var p in m) if (!exports.hasOwnProperty(p)) exports[p] = m[p];
 }
 Object.defineProperty(exports, "__esModule", { value: true });
-const source_1 = require("./source");
-const sink_1 = require("./sink");
-__export(require("./source"));
-__export(require("./sink"));
+__export(require("./consumer"));
+__export(require("./producer"));
 const OFFSET_COMMIT_INTERVAL = 1000;
 const RETENTION_TIME = 1000 * 365 * 24;
 exports.isQuad = (quad) => {
@@ -32,24 +30,3 @@ exports.isProgress = (progress) => {
 const DEFAULT_OPTIONS = {
     concurrency: 8
 };
-const memux = (config) => {
-    const { url, name, input = null, output = null, options = DEFAULT_OPTIONS } = config;
-    if (input == null && output == null) {
-        throw new Error('An input, ouput or both must be provided.');
-    }
-    if (output == null) {
-        return {
-            source: source_1.createSource({ url, name, topic: input })
-        };
-    }
-    if (input == null) {
-        return {
-            sink: sink_1.createSink({ url, name, topic: output, concurrency: options.concurrency })
-        };
-    }
-    return {
-        source: source_1.createSource({ url, name, topic: input }),
-        sink: sink_1.createSink({ url, name, topic: output, concurrency: options.concurrency })
-    };
-};
-exports.default = memux;
