+++ test/producer.js
@@ -1,0 +1,5 @@
+import test from 'ava';
+
+test('stub', t => {
+  t.is(true, true);
+});
